Move journal entries to profile tab
